+++ CHANGELOG.md
@@ -11,10 +11,12 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Added
 - Search for messages by message-Id
 - Search for messages by In-Reply-To
+- Message threading added `Message::thread()`
 - Default folder locations added
 
 ### Affected Classes
 - [Query::class](src/Query/Query.php)
+- [Message::class](src/Message.php)
 
 ### Breaking changes
 -NaN